A chat application to Share and Remind
Worked on a lean start-up model and conceived a novel feature for the
chat application. Added a new 'share and remind' feature (inspired from
Splitwise) along with basic chat. This helps users periodically remind
friends who have borrowed money or things with the help of a customized,
intelligent notification system.
 Frontend: jQuery Mobile; Back-end: PHP.